requirements.txt
flask
gunicorn
psycopg2-binary
sqlalchemy
flask_sqlalchemy
flask_session
beautifulsoup4
selenium
playwright
